requirements.txt
streamlit>=1.35.0
groq>=0.9.0
tensorflow-cpu==2.21.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
plotly>=5.18.0
protobuf>=3.20.2,<7.0.0